trochu už i simulace v.3

+++ Circle.hpp
@@ -33,8 +33,8 @@ public:
     ~Circle() = default;
 private:
     std::string name;
-    int start_x;
-    int start_y;
+    double start_x;
+    double start_y;
     int speed_x;
     int speed_y;
     int creation_time;
